Update billing period in summary
Currently, the billing period is fixed to "ANNUALLY" but we have to make it dynamic to align it with the billing period toggle in the plans-overview. Therefore we need to know the billing period of the selected plan. In this commit we added a selector that retrieves it and assigns it to the right variables, which allows us to retrieve the correct (i.e. "ANNUAL" or "MONTHLY") plans.